+++ ElvUI_MerathilisUI/Core/Changelog/Previous/6.12.lua
@@ -1,15 +1,18 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[612] = {
-	RELEASE_DATE = "TBD",
+	RELEASE_DATE = "25.04.2024",
 	FIXES = {
 		"[TalentLoadoutEx]: Remove skin",
-		"[Details Skin]: Fix gradient Names",
+		"[Details Skin]: Fix gradient names and enable it by default",
 	},
-	NEW = {},
+	NEW = {
+		"[AutoButtons]: Added S4 openable boxes",
+	},
 	IMPROVEMENTS = {
 		"[Raid Manager]: Some code improvements to RaidManager",
 		"[ChatText]: Here also some code improvements",
 		"[RaiderIO Skin]: Updated a new SearchFrame",
+		"[Event Tracker]: Change the defaults on the Map Event Tracker",
 	},
 }