+++ ElvUI_MerathilisUI/Settings/AddonProfiles/FCT.lua
@@ -1,5 +1,5 @@
 local MER, F, E, I, V, P, G, L = unpack(ElvUI_MerathilisUI)
-if not IsAddOnLoaded("ElvUI_FCT") then
+if not C_AddOns.IsAddOnLoaded("ElvUI_FCT") then
 	return
 end