+++ CHANGELOG.md
@@ -3,5 +3,7 @@
   * [Raid Manager]: Some code improvements to RaidManager
   * [ChatText]: Here also some code improvements
   * [RaiderIO Skin]: Updated a new SearchFrame
-  * [Details Skin]: Fix gradient Names
+  * [Details Skin]: Fix gradient Names and enable it by default
   * [TalentLoadoutEx]: Remove skin
+  * [Event Tracker]: Change the defaults on the Map Event Tracker
+  * [AutoButtons]: Added S4 openable boxes